[UPDATE] Se Modifico de lugar  el subir imagenes a su controller y service respectivo y se agrego eliminar imagenes
