+++ src/main/java/mx/edu/utez/carsishop/services/clothes/ClothesService.java
@@ -187,36 +187,5 @@ public class ClothesService {
         );
     }
 
-    public CustomResponse<List<Image>> addImages(ClothesImagesDto clothesImagesDto) throws IOException {
-        Optional<Clothes> clothesOptional = clothesRepository.findById(clothesImagesDto.getClothesId());
-        if(clothesOptional.isEmpty()){
-            return new CustomResponse<>(null, true, 400, "Clothes not found");
-        }
-        if(!clothesImagesDto.isValid()){
-            return new CustomResponse<>(null, true, 400, "Invalid data");
 
-        }
-        List<Image> imagesList = new ArrayList<>();
-        //se suben las imagenes a cloudinary
-        UploadImage uploadImage = new UploadImage();
-        List<ClothesImagesDto.ImagesAndIndex> images = clothesImagesDto.getImages();
-        for (int i = 0; i < images.size(); i++){
-            if(imagesList.size()>=5){
-                break;
-            }
-            Image image = new Image();
-            if(images.get(i).getIndex()>4 || images.get(i).getIndex()<0){
-                continue;
-            }
-            image.setUrl(uploadImage.uploadImage(images.get(i).getImage(), clothesOptional.get().getName()+"-"+images.get(i).getIndex()));
-            image.setClothes(clothesOptional.get());
-            Optional<Image> imageOptional = imageRepository.findByUrl(image.getUrl());
-            if(imageOptional.isEmpty()){
-                imagesList.add(image);
-            }else {
-                imagesList.add(imageOptional.get());
-            }
-        }
-        return  new CustomResponse<>(imageRepository.saveAll(imagesList), false, 200, "Images uploaded");
-    }
 }
